requirements.txt
pyYAML == 6.0.2
pillow == 11.2.1
pytest == 8.4.0 
google-genai == 1.19.0
matplotlib == 3.8.4
pandas == 2.3.1
altair == 5.5.0
python-dotenv == 1.1.1